+++ test_rdflib.py
@@ -1,0 +1,6 @@
+from rdflib import Graph
+from time import time
+start_time = time()
+g = Graph()
+g.parse('http://dbpedia.org/resource/WebAssembly')
+print(f'took {time() - start_time}')
